third_party/blink/web_tests/editing/execCommand/inline-style-after-indentoutdent-expected.txt

This tests createMarkup via insertOrderedList. It makes sure that we don't delete font styles and text decorations. This test must be rewritten when insertOrderedList is modified not to use moveParagraph.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testCreateMarkup('<b>hello</b>') is '<b>hello</b>'
PASS testCreateMarkup('<strong>hello</strong>') is '<strong>hello</strong>'
PASS testCreateMarkup('<i>hello</i>') is '<i>hello</i>'
PASS testCreateMarkup('<em>hello</em>') is '<em>hello</em>'
PASS testCreateMarkup('<s>hello</s>') is '<s>hello</s>'
PASS testCreateMarkup('<strike>hello</strike>') is '<strike>hello</strike>'
PASS testCreateMarkup('<em><s><u>hello</u></s></em>') is '<em><s><u>hello</u></s></em>'
PASS testCreateMarkup('<u><s><em><u>hello</u></em></s></u>') is '<u><s><em><u>hello</u></em></s></u>'
PASS testCreateMarkup('<i><span style="text-decoration-line: underline overline line-through;">world</span></i>') is '<i><span style="text-decoration-line: underline overline line-through;">world</span></i>'
PASS testCreateMarkup('<em><u style="font-weight: bold;">hello</u></em>') is '<em><u style="font-weight: bold;">hello</u></em>'
PASS testCreateMarkup('hello <u>world</u> webkit') is 'hello <u>world</u> webkit'
PASS successfullyParsed is true

TEST COMPLETE
